refactor: relocate module entities into entities/ subfolder
